Fix bad config merge of iframe transport.
